requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
reportlab==4.1.0
sqlparse==0.4.4
typing_extensions==4.11.0
